data_prep/requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine learning libraries
torch>=1.12.0
transformers>=4.20.0

# Chemical informatics
rdkit-pypi>=2022.9.1
pubchempy>=1.0.4

# HTTP requests
requests>=2.28.0

# Parallel processing
concurrent-futures>=3.1.1

# Logging and utilities
pathlib2>=2.3.7 